add fetch lens profiles mechanism
