+++ package.json
@@ -12,7 +12,6 @@
 	},
 	"dependencies": {
 		"@apollo/client": "^3.7.16",
-		"@rainbow-me/rainbowkit": "^1.0.4",
 		"@types/node": "20.3.1",
 		"@types/react": "18.2.13",
 		"@types/react-dom": "18.2.6",
